Include cursor excerpt to the QA prompt
